adding admin controller checks in controller, templates, helper

+++ app/helpers/admin_menu_helper.rb
@@ -30,8 +30,8 @@ module AdminMenuHelper
 
   def admin_dropdown_for_collection(collection, labelled_by_id:)
     options = [
-      link_to('Edit', edit_admin_collection_path(collection), class: "dropdown-item"),
-      link_to('Delete', [:admin, collection], method: :delete, data: { confirm: "Delete collection '#{collection.title}'?" }, class: "dropdown-item")
+      (link_to('Edit', edit_admin_collection_path(collection), class: "dropdown-item") if can?(:update, collection)),
+      (link_to('Delete', [:admin, collection], method: :delete, data: { confirm: "Delete collection '#{collection.title}'?" }, class: "dropdown-item") if can?(:destroy, collection))
     ].compact
 
     content_tag(:div, class: "dropdown-menu dropdown-menu-right", :"aria-labelledby" => labelled_by_id) do
